hoisted props out of contact info view component, moved styles files
