+++ src/components/views/NewContactViewItem/index.tsx
@@ -1,5 +1,5 @@
-import {CredentialRole, IImageAttributes} from '@sphereon/ssi-sdk.data-store';
-import {backgroundColors} from '@sphereon/ui-components.core';
+import {CredentialRole} from '@sphereon/ssi-sdk.data-store';
+import {backgroundColors, ImageAttributes} from '@sphereon/ui-components.core';
 import {useMemo} from 'react';
 import {
   SSIContactViewItemContainerStyled as ContactDetailsHeaderContainer,
@@ -13,18 +13,16 @@ import {SSILogo as Logo} from '@sphereon/ui-components.ssi-react-native';
 import {View} from 'react-native';
 import {VerifiedLabel} from './components/VerifiedLabel';
 
-type NewContactViewItemProps = {
-  // contact: Party;
+export type NewContactViewItemProps = {
   name: string;
-  uri?: string;
-  logo?: IImageAttributes;
+  logo?: ImageAttributes;
   roles: Array<CredentialRole>;
   verified?: boolean;
   background?: 'light' | 'dark';
 };
 
 export const NewContactViewItem = (props: NewContactViewItemProps) => {
-  const {name, uri, roles, logo, verified = false, background = 'dark'} = props;
+  const {name, roles, logo, verified = false, background = 'dark'} = props;
   const backgroundColor = useMemo(() => {
     return background === 'light' ? '#2C334B' : backgroundColors.primaryDark;
   }, [background]);
@@ -37,10 +35,9 @@ export const NewContactViewItem = (props: NewContactViewItemProps) => {
       <View style={{flex: 1, justifyContent: 'center'}}>
         <ContactDetailsContainer>
           <ContactNameCaption>{name}</ContactNameCaption>
-          <ContactRolesCaption>{roles.join(', ')}</ContactRolesCaption>
+          {roles?.length > 0 && <ContactRolesCaption>{roles.join(', ')}</ContactRolesCaption>}
         </ContactDetailsContainer>
         {verified && <VerifiedLabel />}
-        {/* <ContactUriCaption>{uri}</ContactUriCaption> */}
       </View>
     </ContactDetailsHeaderContainer>
   );
