+++ src/styles/components/screens/ContactActivityScreen/index.ts
